update Dockerfile for belua/golang-builder image
This commit updates the Dockerfile to use belua/golang-builder as the base image, which provides a pre-installed set of Go tools, and also upgrades and installs additional dependencies. Specifically, this update adds installation of the spf13/cobra-cli package, upgrades all packages using apt-get upgrade, and installs the wget package.

This change should improve build time and reduce the amount of boilerplate code needed in other projects using this image.