Add latest versions of several Go tools to Dockerfile
This commit updates the Dockerfile to include the latest versions of several Go tools, including gomodifytags, dlv, goplay, impl, errcheck, go-outline, cobra-cli, gopls, workflowcheck, staticcheck, and buf. It also removes the installation of gotests which is not currently needed. This update improves the developer experience by ensuring that developers are using the latest versions of these tools.